% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centenarians_data.R
\docType{data}
\name{centenarians_data}
\alias{centenarians_data}
\title{Centenarians Dataset}
\format{
\subsection{\code{centenarians_data}}{

A data frame with 201 rows and 8 columns:
\describe{
\item{rank}{Rank of the individual}
\item{name}{Name of the individual}
\item{birth_date}{Birth date of the individual}
\item{death_date}{Death date of the individual}
\item{age}{Age of the individual at the time of death or current age if still alive}
\item{place_of_death_or_residence}{Place of death or residence of the individual}
\item{gender}{Gender of the individual}
\item{still_alive}{Status indicating whether the individual is still alive or deceased}
}
}
}
\source{
https://en.wikipedia.org/wiki/List_of_the_verified_oldest_people
}
\usage{
centenarians_data
}
\description{
A dataset of verified oldest people
}
\keyword{datasets}
